Add ingredients fields checkboxes in form pizza edit/create page - create pizzas list as relationship attribute with pizza entity in ingredient model - fix ingredients list in pizza model

+++ src/main/java/org/lessons/springlamiapizzeriacrud/controller/IngredientController.java
@@ -2,12 +2,15 @@ package org.lessons.springlamiapizzeriacrud.controller;
 
 import jakarta.validation.Valid;
 import org.lessons.springlamiapizzeriacrud.model.Ingredient;
+import org.lessons.springlamiapizzeriacrud.model.Pizza;
 import org.lessons.springlamiapizzeriacrud.repository.IngredientRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
 import java.util.Optional;
@@ -65,6 +68,18 @@ public class IngredientController {
 
     @PostMapping("/delete/{id}")
     public String delete(@PathVariable Integer id) {
+        // prima di eliminare l'ingrediente lo dissocio da tutte le pizze
+        Optional<Ingredient> result = ingredientRepository.findById(id);
+        if (result.isEmpty()) {
+            // ritorno un HTTP Status 404 Not Found
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "L'ingrediente con ID " + id + " non è stato trovato");
+        }
+        // ottengo l'oggetto ingrediente da eliminare
+        Ingredient ingredientToDelete = result.get();
+        // itero su ogni pizza associata all'ingrediente per eliminare l'ingrediente
+        for (Pizza pizza : ingredientToDelete.getPizzas()) {
+            pizza.getIngredients().remove(ingredientToDelete);
+        }
         ingredientRepository.deleteById(id);
         return "redirect:/ingredients";
     }
